requirements.txt
# Reddit API
praw==7.7.1

# Data manipulation
pandas==2.1.4
numpy==1.26.2

# Financial data
yfinance==0.2.33

# Sentiment Analysis
vaderSentiment==3.3.2
transformers==4.36.2
torch==2.1.2

# Machine Learning
scikit-learn==1.3.2
xgboost==2.0.3

# Data visualization
matplotlib==3.8.2
seaborn==0.13.0
plotly==5.18.0

# Utilities
python-dotenv==1.0.0
requests==2.31.0
tqdm==4.66.1

# Jupyter
jupyter==1.0.0
ipykernel==6.27.1

# Testing
pytest==7.4.3
pytest-cov==4.1.0
